+++ plugins/catalog-import/src/util/useGithubRepos.ts
@@ -18,7 +18,7 @@ import * as YAML from 'yaml';
 import { useApi, configApiRef } from '@backstage/core';
 import { catalogImportApiRef } from '../api/CatalogImportApi';
 import { ConfigSpec } from '../components/ImportComponentPage';
-import parseGitUri from 'git-url-parse';
+import parseGitUrl from 'git-url-parse';
 
 // TODO: (O5ten) Refactor into a core API instead of direct usage like this
 // https://github.com/backstage/backstage/pull/3613#issuecomment-7408929430
@@ -33,7 +33,7 @@ export function useGithubRepos() {
       name: repoName,
       owner: ownerName,
       resource: hostname,
-    } = parseGitUri(selectedRepo.location);
+    } = parseGitUrl(selectedRepo.location);
 
     const configs = readGitHubIntegrationConfigs(
       config.getOptionalConfigArray('integrations.github') ?? [],