+++ plugins/kubernetes/src/components/KubernetesContent/ErrorPanel.test.tsx
@@ -33,7 +33,7 @@ describe('ErrorPanel', () => {
     // title
     expect(
       getByText(
-        'There was an error retrieving some Kubernetes resources for the entity: THIS_ENTITY',
+        'There was a problem retrieving some Kubernetes resources for the entity: THIS_ENTITY. This could mean that the Error Reporting card is not completely accurate.',
       ),
     ).toBeInTheDocument();
 
@@ -67,7 +67,7 @@ describe('ErrorPanel', () => {
     // title
     expect(
       getByText(
-        'There was an error retrieving some Kubernetes resources for the entity: THIS_ENTITY',
+        'There was a problem retrieving some Kubernetes resources for the entity: THIS_ENTITY. This could mean that the Error Reporting card is not completely accurate.',
       ),
     ).toBeInTheDocument();