+++ plugins/scaffolder-backend/src/scaffolder/stages/prepare/github.ts
@@ -21,7 +21,7 @@ import { parseLocationAnnotation } from '../helpers';
 import { InputError, Git } from '@backstage/backend-common';
 import { PreparerBase, PreparerOptions } from './types';
 import { Logger } from 'winston';
-import GitUriParser from 'git-url-parse';
+import parseGitUrl from 'git-url-parse';
 import { Config } from '@backstage/config';
 import {
   GitHubIntegrationConfig,
@@ -67,7 +67,7 @@ export class GithubPreparer implements PreparerBase {
     }
     const templateId = template.metadata.name;
 
-    const parsedGitLocation = GitUriParser(location);
+    const parsedGitLocation = parseGitUrl(location);
     const repositoryCheckoutUrl = parsedGitLocation.toString('https');
     const tempDir = await fs.promises.mkdtemp(
       path.join(workingDirectory, templateId),