+++ packages/catalog-model/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/catalog-model
 
+## 0.6.1
+
+### Patch Changes
+
+- f3b064e1c: Export the `schemaValidator` helper function.
+- abbee6fff: Implement System, Domain and Resource entity kinds.
+- 147fadcb9: Add subcomponentOf to Component kind to represent subsystems of larger components.
+
 ## 0.6.0
 
 ### Minor Changes