+++ plugins/auth-backend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-auth-backend
 
+## 0.2.10
+
+### Patch Changes
+
+- 468579734: Allow blank certificates and support logout URLs in the SAML provider.
+- Updated dependencies [f3b064e1c]
+- Updated dependencies [abbee6fff]
+- Updated dependencies [147fadcb9]
+  - @backstage/catalog-model@0.6.1
+  - @backstage/backend-common@0.4.3
+
 ## 0.2.9
 
 ### Patch Changes